GHS; Fixed lots of things. Main was a problem in KSTeFile.cpp which caused      the TE file to be corrupted when not using the fortran version of      whipptletilt

+++ common/include/KSTeFile.h
@@ -55,7 +55,7 @@ class KSTeFile
   bool foundReadError(){return fFoundError;};
 
  private:
-  float getMinPeTime(std::vector<KSPixel>& fPixel);
+  double getMinPeTime(std::vector<KSPixel>& fPixel);
 
   std::ifstream* pfInFile;
   std::ofstream* pfOutFile;
